Removed extra script and renamed V2
