+++ bin/create_btk_dataset.py
@@ -1,13 +1,77 @@
 #!/usr/bin/env python3
-"""
-Script for creating a BlobToolKit dataset
+
+VERSION = "2.0.0"
+DESCRIPTION = f"""
+---
+Script for creating a BlobToolKit dataset from the ASCC output files
+Version: {VERSION}
+---
+
+Written by Eerik Aunin (ea10)
+
+Modified by Damon-Lee Pointon (@dp24/@DLBPointon)
+
 """
 
 import general_purpose_functions as gpf
 import argparse
 from pathlib import Path
+import textwrap
 import sys
 import os.path
+
+
+def parse_args(argv=None):
+    parser = argparse.ArgumentParser(
+        prog="createBTKdatasets",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description=textwrap.dedent(DESCRIPTION),
+    )
+    parser.add_argument("-n", "--name", required=True, type=str, help="Assembly name (for the output files)")
+    parser.add_argument(
+        "-tn",
+        "--taxon_name",
+        required=True,
+        type=str,
+        help="The Taxon name of the assembly (Scientific name of the species + subspecies if applicable)",
+    )
+    parser.add_argument("-id", "--taxid", required=True, type=int, help="Taxon ID of the assembly")
+    parser.add_argument(
+        "-td", "--taxdump", required=True, type=str, help="Path to the directory containing the NCBI taxdump"
+    )
+    parser.add_argument("-f", "--fasta", required=True, type=str, help="The path for the assembly fasta file")
+    parser.add_argument(
+        "-d",
+        "--dataset",
+        type=str,
+        required=True,
+        help="The folder containing the data generated throughout the pipeline",
+    )
+    parser.add_argument("-bh", "--blastn_hits", default="N", type=str, help="Path to the BLASTN hits file")
+    parser.add_argument(
+        "-ud", "--uniprot_diamond_hits", default="N", type=str, help="Path to the UNIPROT diamond BlastX hits file"
+    )
+    parser.add_argument("-nr", "--nr_diamond_hits", default="N", type=str, help="Path to the DIAMOND BlastX hits file")
+    parser.add_argument(
+        "-r", "--mapped_reads", default="N", type=str, help="Path to mapped reads BAM for coverage estimation"
+    )
+    parser.add_argument("-t", "--tiara", default="N", type=str, help="Path to the tiara_out.txt file")
+    parser.add_argument(
+        "-p", "--pca", default="N", type=str, help="Path to the kmers_dim_reduction_embeddings.csv file"
+    )
+    parser.add_argument("-fc", "--fcs_gx", default="N", type=str, help="Path to the fcs-gx_summary.csv.csv file")
+    parser.add_argument("-k", "--kraken", default="N", type=str, help="Path to the nt_kraken_lineage.txt file")
+    parser.add_argument("-ms", "--markerscan", default="N", type=str, help="Path to the cobiontid_markerscan.csv file")
+    parser.add_argument("-cv", "--contigviz", default="N", type=str, help="Path to the contigviz_results.csv file")
+    parser.add_argument("-o", "--output", default="btk_datasets", type=str, help="Output directory")
+    parser.add_argument("--threads", type=int, default=1, help="Number of threads to utilise")
+    parser.add_argument("--alias", type=str, default="", help="Assembly alias")
+    parser.add_argument(
+        "--dry_run", dest="dry_run", action="store_true", help="Dry run (print commands without executing)"
+    )
+    parser.add_argument("-v", "--version", action="version", version=VERSION)
+
+    return parser.parse_args(argv)
 
 
 def create_assembly_yaml(assembly_yaml_path, assembly_alias, taxon_name):
@@ -65,167 +129,76 @@ def detect_dim_reduction_methods(kmers_dim_reduction_output_path):
     return dim_reduction_methods
 
 
-def add_custom_variables_to_btk_dataset(pipeline_run_folder, btk_dataset_folder):
-    """
-    Script for adding custom variables (e.g. Tiara results and PCA results) to the BlobToolKit dataset
-    """
-    pipeline_output_folder = pipeline_run_folder + "/collected_tables"
-    if os.path.isdir(pipeline_output_folder) == False:
-        sys.stderr.write(
-            "The directory for the output tables of the pipeline ({}) was not found\n".format(pipeline_output_folder)
-        )
-        sys.exit(1)
-    if os.path.isdir(btk_dataset_folder) == False:
-        sys.stderr.write("The BlobToolKit dataset directory ({}) was not found\n".format(btk_dataset_folder))
-        sys.exit(1)
-    tiara_raw_output_path = pipeline_output_folder + "/tiara_out.txt"
-    if os.path.isfile(tiara_raw_output_path) and os.stat(tiara_raw_output_path).st_size > 0:
-        tiara_reformatted_output_path = pipeline_output_folder + "/tiara_out_btk_format.tsv"
-        tiara_results_to_btk_format(tiara_raw_output_path, tiara_reformatted_output_path)
-        add_tiara_command = 'blobtools add --text {} --text-delimiter "\t" --text-cols "identifier=identifiers,tiara=tiara" --text-header {}'.format(
-            tiara_reformatted_output_path, btk_dataset_folder
-        )
-        gpf.run_system_command(add_tiara_command)
+def main(args):
+    command_list = []
 
-    kmers_dim_reduction_output_path = pipeline_output_folder + "/kmers_dim_reduction_embeddings.csv"
-    if os.path.isfile(kmers_dim_reduction_output_path) and os.stat(kmers_dim_reduction_output_path).st_size > 0:
-        used_dim_reduction_methods = detect_dim_reduction_methods(kmers_dim_reduction_output_path)
-        for dim_reduction_method in used_dim_reduction_methods:
-            add_embedding_command = 'blobtools add --text {path} --text-delimiter "," --text-cols scaff=identifiers,embedding_x_{dim_reduction_method}=embedding_x_{dim_reduction_method},embedding_y_{dim_reduction_method}=embedding_y_{dim_reduction_method} --text-header {btk_dataset_folder}'.format(
-                path=kmers_dim_reduction_output_path,
-                dim_reduction_method=dim_reduction_method,
-                btk_dataset_folder=btk_dataset_folder,
-            )
-            gpf.run_system_command(add_embedding_command)
+    assembly_alias = args.name if args.alias == "" else args.alias
 
-    kraken_lineage_path = pipeline_output_folder + "/nt_kraken_lineage.txt"
-    if os.path.isfile(kraken_lineage_path) and os.stat(kraken_lineage_path).st_size > 0:
-        for taxonomy_level in ("species", "genus", "family", "order", "class", "phylum", "kingdom", "domain"):
-            add_kraken_command = 'blobtools add --text {} --text-delimiter "," --text-cols scaff=identifiers,nt_kraken_{}=nt_kraken_{} --text-header {}'.format(
-                kraken_lineage_path, taxonomy_level, taxonomy_level, btk_dataset_folder
-            )
-            gpf.run_system_command(add_kraken_command)
+    edited_assembly_title = args.name.replace(".", "_").replace(" ", "_")
 
-    fcs_gx_output_path = pipeline_output_folder + "/fcs-gx_summary.csv"
-    if os.path.isfile(fcs_gx_output_path) and os.stat(fcs_gx_output_path).st_size > 0:
-        add_fcs_gx_results_command = 'blobtools add --text {} --text-delimiter "," --text-cols "scaff=identifiers,fcs_gx_top_tax_name=fcs_gx_top_tax_name,fcs_gx_div=fcs_gx_div,fcs_gx_action=fcs_gx_action" --text-header {}'.format(
-            fcs_gx_output_path, btk_dataset_folder
-        )
-        gpf.run_system_command(add_fcs_gx_results_command)
+    assembly_yaml_path = args.output + "/" + edited_assembly_title + "BTK_DS.yaml"
 
-    # cobiontid_markerscan_json_file_path = run_folder + "/" + sample_id + ".json"
-    # cobiontid_scaffs_json_to_csv(json_file_path, out_folder + "/cobiontid_markerscan.csv")
-    cobiontid_markerscan_output_path = pipeline_output_folder + "/cobiontid_markerscan.csv"
-    if os.path.isfile(cobiontid_markerscan_output_path) and os.stat(cobiontid_markerscan_output_path).st_size > 0:
-        add_cobiontid_markerscan_results_command = 'blobtools add --text {} --text-delimiter "," --text-cols "scaff=identifiers,CobiontID_MarkerScan_embl_ebi_ena=CobiontID_MarkerScan_embl_ebi_ena,CobiontID_MarkerScan_slv=CobiontID_MarkerScan_slv,CobiontID_MarkerScan_Cluster=CobiontID_MarkerScan_Cluster" --text-header {}'.format(
-            cobiontid_markerscan_output_path, btk_dataset_folder
-        )
-        gpf.run_system_command(add_cobiontid_markerscan_results_command)
+    if args.dry_run == False:
+        Path(args.dataset).mkdir(parents=True, exist_ok=True)
+        create_assembly_yaml(assembly_yaml_path, assembly_alias, args.taxon_name)
 
-    cobiontid_contigviz_output_path = pipeline_output_folder + "/contigviz_results.csv"
-    if os.path.isfile(cobiontid_contigviz_output_path) and os.stat(cobiontid_contigviz_output_path).st_size > 0:
-        add_cobiontid_contigviz_results_command = 'blobtools add --text {} --text-delimiter "," --text-cols "scaff=identifiers,ContigViz_UMAP1=ContigViz_UMAP1,ContigViz_UMAP2=ContigViz_UMAP2,ContigViz_Hexamer_continuous=ContigViz_Hexamer_continuous,ContigViz_Hexamer_digitized=ContigViz_Hexamer_digitized,ContigViz_FastK_continuous=ContigViz_FastK_continuous,ContigViz_FastK_digitized=ContigViz_FastK_digitized,ContigViz_Unique_15mers_continuous=ContigViz_Unique_15mers_continuous,ContigViz_Unique_15mers_digitized=ContigViz_Unique_15mers_digitized,ContigViz_Coverage_continuous=ContigViz_Coverage_continuous,ContigViz_Coverage_digitized=ContigViz_Coverage_digitized" --text-header {}'.format(
-            cobiontid_contigviz_output_path, btk_dataset_folder
-        )
-        gpf.run_system_command(add_cobiontid_contigviz_results_command)
+    # Base command for new BTK Dataset
+    blobtools_create_command = f"blobtools create --fasta {args.fasta} --meta {assembly_yaml_path} --taxid {args.taxid} --taxdump {args.taxdump} {args.output}"
+    gpf.run_system_command(blobtools_create_command, dry_run=args.dry_run)
 
+    # ADDING BLAST HIT DATA TO BTK
+    hits_file_paths = [args.blastn_hits, args.uniprot_diamond_hits, args.nr_diamond_hits]
 
-def main(
-    assembly_fasta_path,
-    dataset_folder,
-    pipeline_run_folder,
-    assembly_title,
-    taxon_name,
-    taxid,
-    blastn_hits_path,
-    uniprot_diamond_hits_path,
-    nr_diamond_hits_path,
-    mapped_reads_path,
-    taxdump_path,
-    threads,
-    assembly_alias,
-    dry_run_flag,
-):
-    # out_folder = pipeline_run_folder + "/collected_tables"
+    hits_file = [n for n in hits_file_paths if n != "N" and os.path.isfile(n) is True and os.stat(n).st_size > 0]
 
-    if assembly_alias == "":
-        assembly_alias = assembly_title
-
-    if dry_run_flag == False:
-        Path(dataset_folder).mkdir(parents=True, exist_ok=True)
-
-    edited_assembly_title = assembly_title.replace(".", "_")
-    edited_assembly_title = edited_assembly_title.replace(" ", "_")
-
-    assembly_yaml_path = dataset_folder + "/" + edited_assembly_title + ".yaml"
-    if dry_run_flag == False:
-        create_assembly_yaml(assembly_yaml_path, assembly_alias, taxon_name)
-
-    blobtools_create_command = "blobtools create --fasta {} --meta {} --taxid {} --taxdump {} {}".format(
-        assembly_fasta_path, assembly_yaml_path, taxid, taxdump_path, dataset_folder
-    )
-    gpf.run_system_command(blobtools_create_command, dry_run=dry_run_flag)
-
-    hits_file_paths = [blastn_hits_path, uniprot_diamond_hits_path, nr_diamond_hits_path]
-    hits_file_paths = [n for n in hits_file_paths if os.path.isfile(n) is True and os.stat(n).st_size > 0]
-
-    if len(hits_file_paths) > 0:
+    if len(hits_file) > 0:
         add_hits_command = "blobtools add"
-        for hits_file_path in hits_file_paths:
-            add_hits_command += " --hits {}".format(hits_file_path)
-        add_hits_command += " --taxrule bestsum --taxdump {} {}".format(taxdump_path, dataset_folder)
-        gpf.run_system_command(add_hits_command, dry_run=dry_run_flag)
+        for file in hits_file_paths:
+            add_hits_command += f" --hits {file}"
+        add_hits_command += f" --taxrule bestsum --taxdump {args.taxdump} {args.output}"
+        command_list.append(add_hits_command)
 
-    if os.path.isfile(mapped_reads_path) is True and os.stat(mapped_reads_path).st_size > 0:
-        add_cov_command = "blobtools add --cov {} --threads {} {}".format(mapped_reads_path, threads, dataset_folder)
-        gpf.run_system_command(add_cov_command, dry_run=dry_run_flag)
+    # ADDING MAPPED READS DATA TO BTK
+    if (
+        args.mapped_reads != "N"
+        and os.path.isfile(args.mapped_reads) is True
+        and os.stat(args.mapped_reads).st_size > 0
+    ):
+        add_cov_command = f"blobtools add --cov {args.mapped_reads} --threads {args.threads} {args.output}"
+        command_list.append(add_cov_command)
 
-    # export_table_command = "blobtools filter --table {}/btk_summary_table_basic.tsv {}".format(out_folder, dataset_folder)
-    add_custom_variables_to_btk_dataset(pipeline_run_folder, dataset_folder)
-    export_table_command = "blobtools filter --table {}/collected_tables/btk_summary_table_full.tsv {}".format(
-        pipeline_run_folder, dataset_folder
-    )
+    # ADDING TIARA
+    if args.tiara != "N" and os.path.isfile(args.tiara) and os.stat(args.tiara).st_size > 0:
+        tiara_reformatted_output_path = args.dataset + "/tiara_out_btk_format.tsv"
+        tiara_results_to_btk_format(args.tiara, tiara_reformatted_output_path)
+        add_tiara_command = f"blobtools add --text {tiara_reformatted_output_path} --text-delimiter '\t' --text-cols 'identifier=identifiers,tiara=tiara' --text-header {args.output}"
+        command_list.append(add_tiara_command)
 
-    gpf.run_system_command(export_table_command, dry_run=dry_run_flag)
+    # ADDING KMER DIM REDUCTION
+    if args.pca != "N" and os.path.isfile(args.pca) and os.stat(args.pca).st_size > 0:
+        used_dim_reduction_methods = detect_dim_reduction_methods(args.pca)
+        for dim_reduction_method in used_dim_reduction_methods:
+            add_embedding_command = f"blobtools add --text {args.pca} --text-delimiter ',' --text-cols scaff=identifiers,embedding_x_{dim_reduction_method}=embedding_x_{dim_reduction_method},embedding_y_{dim_reduction_method}=embedding_y_{dim_reduction_method} --text-header {args.output}"
+            command_list.append(add_embedding_command)
 
-    # json_file_path = run_folder + "/" + sample_id + ".json"
-    # cobiontid_scaffs_json_to_csv(json_file_path, out_folder + "/cobiontid_markerscan.csv")
+    # ADDIND KRAKEN DATA
+    if args.kraken != "N" and os.path.isfile(args.kraken) and os.stat(args.kraken).st_size > 0:
+        for taxonomy_level in ("species", "genus", "family", "order", "class", "phylum", "kingdom", "domain"):
+            add_kraken_command = f"blobtools add --text {args.kraken} --text-delimiter ',' --text-cols scaff=identifiers,nt_kraken_{taxonomy_level}=nt_kraken_{taxonomy_level} --text-header {args.output}"
+            command_list.append(add_kraken_command)
+
+    # ADDING FCS_GX DATA
+    if args.fcs_gx != "N" and os.path.isfile(args.fcs_gx) and os.stat(args.fcs_gx).st_size > 0:
+        add_fcs_gx_results_command = f"blobtools add --text {args.fcs_gx} --text-delimiter ',' --text-cols 'scaff=identifiers,fcs_gx_top_tax_name=fcs_gx_top_tax_name,fcs_gx_div=fcs_gx_div,fcs_gx_action=fcs_gx_action' --text-header {args.output}"
+        command_list.append(add_fcs_gx_results_command)
+
+    export_table_command = f"blobtools filter --table btk_summary_table_full.tsv {args.output}"
+    command_list.append(export_table_command)
+
+    # EXECUTE ALL BTK COMMANDS
+    for i in command_list:
+        gpf.run_system_command(i, dry_run=args.dry_run)
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("assembly_fasta_path", type=str, help="assembly_fasta_path")
-    parser.add_argument("dataset_folder", type=str, help="Path for dataset folder")
-    parser.add_argument("pipeline_run_folder", type=str, help="Folder where this pipeline is run pipeline")
-    parser.add_argument("assembly_title", type=str, help="Assembly title")
-    parser.add_argument("-tn", "--taxon_name", type=str, help="Taxon name")
-    parser.add_argument("-ti", "--taxid", type=int, help="taxid")
-    parser.add_argument("blastn_hits_path", type=str, help="Path to blastn hits file")
-    parser.add_argument("uniprot_diamond_hits_path", type=str, help="Path to UNIPROT Diamond BLASTX hits file")
-    parser.add_argument("nr_diamond_hits_path", type=str, help="Path to nr Diamond BLASTX hits file")
-    parser.add_argument(
-        "mapped_reads_path", type=str, help="Path to the BAM file with mapped reads for coverage estimation"
-    )
-    parser.add_argument("-td", "--taxdump_path", type=str, help="Path to the directory with NCBI taxdump files")
-    parser.add_argument("--threads", type=int, default=1, help="Number of CPU threads (default: 1)")
-    parser.add_argument("--assembly_alias", type=str, default="", help="Assembly alias")
-    parser.add_argument(
-        "--dry_run", dest="dry_run", action="store_true", help="Dry run (print commands without executing)"
-    )
-    args = parser.parse_args()
-    main(
-        args.assembly_fasta_path,
-        args.dataset_folder,
-        args.pipeline_run_folder,
-        args.assembly_title,
-        args.taxon_name,
-        args.taxid,
-        args.blastn_hits_path,
-        args.uniprot_diamond_hits_path,
-        args.nr_diamond_hits_path,
-        args.mapped_reads_path,
-        args.taxdump_path,
-        args.threads,
-        args.assembly_alias,
-        args.dry_run,
-    )
+    main(parse_args())
